+++ lib/mhgu_query/models/item.rb
@@ -1,0 +1,8 @@
+require_relative '../db'
+
+module MHGUQuery
+  module Models
+    class Item < Sequel::Model
+    end
+  end
+end